pyproject.toml
[tool.poetry]
name = "cdk"
version = "0.1.0"
description = ""
authors = ["komakichi <kumaou003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cdk-aws-lambda-powertools-layer = "^3.8.0"
constructs = "^10.3.0"
scrapy = "^2.11.2"
slack-sdk = "^3.33.1"
gspread = "^6.1.2"
oauth2client = "^4.1.3"
google-api-python-client = "^2.147.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[tool.poetry.group.dev.dependencies]
aws-lambda-powertools = "^3.0.0"
aws-cdk-lib = "^2.160.0"
aws-cdk-aws-lambda-python-alpha = "^2.160.0a0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
